Feat: Modularize application
